pyproject.toml
[project]
name = "cris"
version = "0.1.0"
description = "Criminal Reasoning Intelligence System - Multi-agent AI platform for criminal intelligence and investigation"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "CRIS Contributors" }]
keywords = ["criminal-intelligence", "multi-agent", "ai", "investigation", "neo4j", "knowledge-graph"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Legal Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Web Framework
    "streamlit>=1.31.0",
    
    # Database
    "neo4j>=5.17.0",
    "chromadb>=0.4.22",
    
    # LLM & AI
    "google-generativeai>=0.4.0",
    "openai>=1.12.0",
    "anthropic>=0.18.0",
    "sentence-transformers>=2.3.0",
    
    # Data Validation & Settings
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "python-dotenv>=1.0.0",
    
    # Data Processing
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    
    # Geospatial
    "folium>=0.15.0",
    "geopandas>=0.14.0",
    "streamlit-folium>=0.18.0",
    "geopy>=2.4.0",
    
    # Document Processing
    "pypdf2>=3.0.0",
    "pytesseract>=0.3.10",
    "pillow>=10.2.0",
    
    # Graph Visualization
    "networkx>=3.2.0",
    "pyvis>=0.3.2",
    
    # Charts & Visualization
    "plotly>=5.18.0",
    
    # HTTP & Async
    "httpx>=0.26.0",
    "tenacity>=8.2.0",
    
    # Logging
    "structlog>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.1.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.24.0",
]

[project.urls]
Homepage = "https://github.com/cris-project/cris"
Documentation = "https://cris-project.github.io/cris"
Repository = "https://github.com/cris-project/cris"
Issues = "https://github.com/cris-project/cris/issues"

[project.scripts]
cris = "cris.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "B905",   # zip without explicit strict
]

[tool.ruff.lint.isort]
known-first-party = ["config", "core", "database", "agents", "processors", "models", "ui", "services", "utils"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=. --cov-report=term-missing"
